fix: Splash 렌더링 방식 수정 및 import 관련 ESLint 임시 비활성화
