+++ eslint.config.mjs
@@ -47,32 +47,32 @@ const eslintConfig = [
 					tsx: 'never',
 				},
 			],
-			'import/order': [
-				'error',
-				{
-					groups: [['builtin', 'external', 'internal'], ['parent', 'sibling', 'index'], 'unknown'],
-					pathGroups: [
-						{ pattern: 'next*', group: 'builtin', position: 'before' },
-						{ pattern: 'next*/**', group: 'builtin', position: 'before' },
-						{ pattern: 'react*', group: 'builtin', position: 'before' },
-						{ pattern: 'react*/**', group: 'builtin', position: 'before' },
-						{ pattern: '@/services/**', group: 'internal', position: 'after' },
-						{ pattern: '@/lib/**', group: 'internal', position: 'after' },
-						{ pattern: '@/hooks/**', group: 'internal', position: 'after' },
-						{ pattern: '@/components/**', group: 'internal', position: 'after' },
-						{ pattern: '@/types/**', group: 'internal', position: 'after' },
-						{ pattern: '@/**', group: 'internal', position: 'after' },
-						{ pattern: '../**', group: 'internal', position: 'after' },
-						{ pattern: './**', group: 'internal', position: 'after' },
-					],
-					pathGroupsExcludedImportTypes: [],
-					alphabetize: {
-						order: 'asc',
-						caseInsensitive: true,
-					},
-					'newlines-between': 'always',
-				},
-			],
+			// 'import/order': [
+			// 	'error',
+			// 	{
+			// 		groups: [['builtin', 'external', 'internal'], ['parent', 'sibling', 'index'], 'unknown'],
+			// 		pathGroups: [
+			// 			{ pattern: 'next*', group: 'builtin', position: 'before' },
+			// 			{ pattern: 'next*/**', group: 'builtin', position: 'before' },
+			// 			{ pattern: 'react*', group: 'builtin', position: 'before' },
+			// 			{ pattern: 'react*/**', group: 'builtin', position: 'before' },
+			// 			{ pattern: '@/services/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '@/lib/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '@/hooks/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '@/components/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '@/types/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '@/**', group: 'internal', position: 'after' },
+			// 			{ pattern: '../**', group: 'internal', position: 'after' },
+			// 			{ pattern: './**', group: 'internal', position: 'after' },
+			// 		],
+			// 		pathGroupsExcludedImportTypes: [],
+			// 		alphabetize: {
+			// 			order: 'asc',
+			// 			caseInsensitive: true,
+			// 		},
+			// 		'newlines-between': 'always',
+			// 	},
+			// ],
 		},
 	},
 ];
